move auth to correct location. Fix bug in tsup config
