+++ packages/server/package.json
@@ -45,7 +45,7 @@
         "cross-env": "^7.0.3",
         "nodemon": "^3.1.0",
         "ts-node": "^10.9.2",
-        "tsup": "^8.0.2",
+        "tsup": "^8.3.0",
         "vitest": "^1.4.0"
     },
     "packageManager": "pnpm@8.15.6+sha256.01c01eeb990e379b31ef19c03e9d06a14afa5250b82e81303f88721c99ff2e6f"
