+++ packages/server/src/auth.ts
